route file naming according to v2 convention
