+++ app/routes/admin.delete.tsx
