+++ app/routes/admin.new.tsx
@@ -199,137 +199,141 @@ const NewContent = () => {
   }, [actionData]);
 
   return (
-    <Form
-      method="post"
-      className="w-full rounded-3xl bg-variant-blue p-5 text-left text-white"
-    >
-      <Input
-        error={actionData?.errors?.title}
-        label="Tittel"
-        htmlRef={titleRef}
-        name="title"
-      />
-
-      <TextArea
-        error={actionData?.errors?.description}
-        label="Beskrivelse"
-        htmlRef={descriptionRef}
-        name="description"
-      />
-
-      <Input
-        error={actionData?.errors?.url}
-        label="Url"
-        htmlRef={urlRef}
-        name="url"
-      />
-
-      <Input
-        error={actionData?.errors?.image}
-        label={
-          <span>
-            Base64-enkodet bilde: Kan genereres{" "}
-            <a
-              href="https://www.base64-image.de/"
-              className="text-variant-pink"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              HER
-            </a>
-          </span>
-        }
-        htmlRef={imageRef}
-        name="image"
-      />
-
-      <Input
-        error={actionData?.errors?.imageText}
-        label="Alternativ tekst for bilde"
-        htmlRef={imageAltTextRef}
-        name="imageText"
-      />
-
-      <Input
-        error={actionData?.errors?.author}
-        label="Forfatter (Variant-epost)"
-        htmlRef={authorRef}
-        name="author"
-      />
-
-      <div className="mt-3 grid grid-cols-2 pb-4 md:grid-cols-3">
-        <Toggle
-          leftText="Nei"
-          rightText="Ja"
-          label="Fremhevet"
-          inputName="featured"
+    <div className="items-left mt-5 flex justify-center">
+      <Form
+        method="post"
+        className="w-full rounded-3xl bg-variant-blue p-5 text-left text-white"
+      >
+        <Input
+          error={actionData?.errors?.title}
+          label="Tittel"
+          htmlRef={titleRef}
+          name="title"
         />
 
-        <fieldset className="sm:mt-4 md:col-span-2">
-          <legend>Kategori:</legend>
-          <div className="mt-3 gap-4 md:flex">
-            {categories?.map((category) => {
-              return (
-                <div key={category.id} className="flex items-center gap-1">
-                  <input
-                    className="h-4 w-4 cursor-pointer"
-                    type="radio"
-                    ref={categoriesRef}
-                    name="categoryId"
-                    id="category"
-                    value={category.id}
-                  />
-                  <label className="font-bold" htmlFor="category">
-                    {category.text}
-                  </label>
-                </div>
-              );
-            })}
-          </div>
+        <TextArea
+          error={actionData?.errors?.description}
+          label="Beskrivelse"
+          htmlRef={descriptionRef}
+          name="description"
+        />
 
-          {actionData?.errors?.categoryId && (
-            <div className="pb-1 text-variant-pink-2" id="error">
-              {actionData?.errors?.categoryId}
+        <Input
+          error={actionData?.errors?.url}
+          label="Url"
+          htmlRef={urlRef}
+          name="url"
+        />
+
+        <Input
+          error={actionData?.errors?.image}
+          label={
+            <span>
+              Base64-enkodet bilde: Kan genereres{" "}
+              <a
+                href="https://www.base64-image.de/"
+                className="text-variant-pink"
+                target="_blank"
+                rel="noopener noreferrer"
+              >
+                HER
+              </a>
+            </span>
+          }
+          htmlRef={imageRef}
+          name="image"
+        />
+
+        <Input
+          error={actionData?.errors?.imageText}
+          label="Alternativ tekst for bilde"
+          htmlRef={imageAltTextRef}
+          name="imageText"
+        />
+
+        <Input
+          error={actionData?.errors?.author}
+          label="Forfatter (Variant-epost)"
+          htmlRef={authorRef}
+          name="author"
+        />
+
+        <div className="mt-3 grid grid-cols-2 pb-4 md:grid-cols-3">
+          <Toggle
+            leftText="Nei"
+            rightText="Ja"
+            label="Fremhevet"
+            inputName="featured"
+          />
+
+          <fieldset className="sm:mt-4 md:col-span-2">
+            <legend>Kategori:</legend>
+            <div className="mt-3 gap-4 md:flex">
+              {categories?.map((category) => {
+                return (
+                  <div key={category.id} className="flex items-center gap-1">
+                    <input
+                      className="h-4 w-4 cursor-pointer"
+                      type="radio"
+                      ref={categoriesRef}
+                      name="categoryId"
+                      id="category"
+                      value={category.id}
+                    />
+                    <label className="font-bold" htmlFor="category">
+                      {category.text}
+                    </label>
+                  </div>
+                );
+              })}
             </div>
-          )}
+
+            {actionData?.errors?.categoryId && (
+              <div className="pb-1 text-variant-pink-2" id="error">
+                {actionData?.errors?.categoryId}
+              </div>
+            )}
+          </fieldset>
+        </div>
+
+        <fieldset className="mt-3 grid grid-cols-2 gap-4 sm:grid-cols-5">
+          <legend>Tags:</legend>
+          {tags?.map((tag) => {
+            return (
+              <div key={tag.id} className="flex items-center gap-1 ">
+                <input
+                  className="h-4 w-4 cursor-pointer"
+                  type="checkbox"
+                  ref={tagsRef}
+                  name="tag"
+                  id={tag.id}
+                  value={tag.id}
+                />
+                <label className="font-bold" htmlFor="tag">
+                  {tag.text}
+                </label>
+              </div>
+            );
+          })}
         </fieldset>
-      </div>
 
-      <fieldset className="mt-3 grid grid-cols-2 gap-4 sm:grid-cols-5">
-        <legend>Tags:</legend>
-        {tags?.map((tag) => {
-          return (
-            <div key={tag.id} className="flex items-center gap-1 ">
-              <input
-                className="h-4 w-4 cursor-pointer"
-                type="checkbox"
-                ref={tagsRef}
-                name="tag"
-                id={tag.id}
-                value={tag.id}
-              />
-              <label className="font-bold" htmlFor="tag">
-                {tag.text}
-              </label>
-            </div>
-          );
-        })}
-      </fieldset>
-
-      <div className="mt-5 flex justify-end">
-        <PrimaryButton
-          type="submit"
-          text={
-            navigation.state === "submitting" || navigation.state === "loading"
-              ? "Lagrer ... "
-              : "Lagre"
-          }
-          disabled={
-            navigation.state === "submitting" || navigation.state === "loading"
-          }
-        />
-      </div>
-    </Form>
+        <div className="mt-5 flex justify-end">
+          <PrimaryButton
+            type="submit"
+            text={
+              navigation.state === "submitting" ||
+              navigation.state === "loading"
+                ? "Lagrer ... "
+                : "Lagre"
+            }
+            disabled={
+              navigation.state === "submitting" ||
+              navigation.state === "loading"
+            }
+          />
+        </div>
+      </Form>
+    </div>
   );
 };
