--- app/routes/admin/content.tsx
@@ -1,11 +1,0 @@
-import { Outlet } from "@remix-run/react";
-
-const Content = () => {
-  return (
-    <div className="items-left mt-5 flex justify-center">
-      <Outlet />
-    </div>
-  );
-};
-
-export default Content;
