+++ app/routes/courses.tsx
@@ -17,6 +17,7 @@ const Course = () => {
       <section className="max-w-5xl">
         <div className="grid gap-4 sm:grid-cols-1 md:grid-cols-2 lg:grid-cols-3">
           {courses?.map((course) => {
+            console.log("course", course);
             return (
               <div key={course.id} className="my-1 inline-grid">
                 <CourseCard
@@ -31,16 +32,6 @@ const Course = () => {
               </div>
             );
           })}
-          <CourseCard
-            title="Intro til HTML"
-            linkToCourse=""
-            image="/assets/icons/course.svg"
-            createdAt="11.06.2023"
-            altImageText="Bildeteskt"
-            description="Beskrivelse her av en beginner app"
-            id=""
-            tags={["beginner"]}
-          />
         </div>
       </section>
     </main>
