+++ remix.config.js
@@ -4,4 +4,7 @@
 module.exports = {
   cacheDirectory: "./node_modules/.cache/remix",
   ignoredRouteFiles: ["**/.*", "**/*.css", "**/*.test.{js,jsx,ts,tsx}"],
+  future: {
+    v2_routeConvention: true,
+  },
 };
